refactor: moved api code to typescript
