+++ api/db/getTop10Records.ts
