+++ api/db/index.ts
@@ -1,15 +1,18 @@
 import { Pool } from 'pg'
 
-let pool = null
+let pool: Pool | null = null
 
 export function getPool() {
-  if (!pool || !pool.connected) {
+  if (!process.env.DB_USER || !process.env.DB_HOST || !process.env.DB_NAME || !process.env.DB_PASSWORD || !process.env.DB_PORT) {
+    throw new Error('Missing database environment variables')
+  }
+  if (!pool) {
     pool = new Pool({
       user: process.env.DB_USER,
       host: process.env.DB_HOST,
       database: process.env.DB_NAME,
       password: process.env.DB_PASSWORD,
-      port: process.env.DB_PORT,
+      port: parseInt(process.env.DB_PORT),
       max: 5, // Maximum number of connections in the pool
       idleTimeoutMillis: 30000, // Close idle connections after 30 seconds
       connectionTimeoutMillis: 2000, // How long to wait for a connection from the pool
