+++ api/types.ts
@@ -1,0 +1,7 @@
+export interface DBRecord {
+  data: string
+  id: string
+  creationTimestamp: string
+}
+
+export type DBInsertRecord = Omit<DBRecord, 'creationTimestamp'>
